Refactor decouple execution (#84)
* Refactor cairovm execution engine

* Refactor cairo native engine
