+++ sequencer/Cargo.toml
@@ -1,5 +1,6 @@
 [workspace]
 members = [
+    "execution_engine",
     "rpc_endpoint",
     "store", 
     "crypto",
