+++ sequencer/execution_engine/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "execution_engine"
+authors = ["LambdaClass <hello@lambdaclass.com>"]
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+cairo-vm = { version = "0.8.1", features = ["cairo-1-hints"]}
+cairo-lang-starknet = { version = "2.0.2", default-features = false }
+serde_json = "1.0.61"
+cairo-native = { git = "https://github.com/lambdaclass/cairo_native", rev="9433c7e"}
+cairo-lang-sierra = "2.0.2"
+num-bigint = "0.4.3"
+cairo-lang-compiler = "2.0.2"
